+++ components/snippets/BlogSearch.js
@@ -1,0 +1,16 @@
+import styled, { css } from 'styled-components';
+import * as m from '@/styles/_mixins';
+
+export const BlogSearch = styled.div`
+  display: flex;
+  padding-top: 2.5rem;
+  justify-content: space-between;
+  flex-direction: column-reverse;
+  align-items: flex-start;
+
+  //media query mixins
+  ${m.desktop(css`
+    align-items: center;
+    flex-direction: row;
+  `)}
+`;